+++ config/configmapprovider/env.go
@@ -37,20 +37,18 @@ func NewEnv() Provider {
 	return &envMapProvider{}
 }
 
-func (emp *envMapProvider) Retrieve(_ context.Context, location string, _ WatcherFunc) (Retrieved, error) {
+func (emp *envMapProvider) Retrieve(_ context.Context, location string, _ WatcherFunc) (*config.Map, EndWatchFunc, error) {
 	if !strings.HasPrefix(location, envSchemeName+":") {
-		return nil, fmt.Errorf("%v location is not supported by %v provider", location, envSchemeName)
+		return nil, nil, fmt.Errorf("%v location is not supported by %v provider", location, envSchemeName)
 	}
 
 	content := os.Getenv(location[len(envSchemeName)+1:])
 	var data map[string]interface{}
 	if err := yaml.Unmarshal([]byte(content), &data); err != nil {
-		return nil, fmt.Errorf("unable to parse yaml: %w", err)
+		return nil, nil, fmt.Errorf("unable to parse yaml: %w", err)
 	}
 
-	return NewRetrieved(func(ctx context.Context) (*config.Map, error) {
-		return config.NewMapFromStringMap(data), nil
-	})
+	return config.NewMapFromStringMap(data), nil, nil
 }
 
 func (*envMapProvider) Shutdown(context.Context) error {