+++ service/config_provider_test.go
@@ -34,20 +34,26 @@ import (
 )
 
 type mockProvider struct {
-	ret  *fakeRetrieved
+	retM *config.Map
+	errW error
+	errC error
 	errR error
 	errS error
 }
 
-func (m *mockProvider) Retrieve(_ context.Context, _ string, watcher configmapprovider.WatcherFunc) (configmapprovider.Retrieved, error) {
+func (m *mockProvider) Retrieve(_ context.Context, _ string, watcher configmapprovider.WatcherFunc) (*config.Map, configmapprovider.EndWatchFunc, error) {
 	if m.errR != nil {
-		return nil, m.errR
+		return nil, nil, m.errR
 	}
-	if m.ret == nil {
-		return &fakeRetrieved{}, nil
+	if m.retM == nil {
+		return config.NewMap(), nil, nil
 	}
-	m.ret.watcher = watcher
-	return m.ret, nil
+
+	if watcher != nil {
+		watcher(&configmapprovider.ChangeEvent{Error: m.errW})
+	}
+
+	return m.retM, func(ctx context.Context) error { return m.errC }, nil
 }
 
 func (m *mockProvider) Scheme() string {
@@ -66,32 +72,6 @@ func (ecu *errConfigUnmarshaler) Unmarshal(*config.Map, component.Factories) (*c
 	return nil, ecu.err
 }
 
-type fakeRetrieved struct {
-	configmapprovider.Retrieved
-	retM    *config.Map
-	errG    error
-	errW    error
-	errC    error
-	watcher configmapprovider.WatcherFunc
-}
-
-func (er *fakeRetrieved) Get(context.Context) (*config.Map, error) {
-	if er.watcher != nil {
-		er.watcher(&configmapprovider.ChangeEvent{Error: er.errW})
-	}
-	if er.errG != nil {
-		return nil, er.errG
-	}
-	if er.retM == nil {
-		return config.NewMap(), nil
-	}
-	return er.retM, nil
-}
-
-func (er *fakeRetrieved) Close(context.Context) error {
-	return er.errC
-}
-
 func TestConfigProvider_Errors(t *testing.T) {
 	factories, errF := componenttest.NopFactories()
 	require.NoError(t, errF)
@@ -107,7 +87,7 @@ func TestConfigProvider_Errors(t *testing.T) {
 		expectShutdownErr bool
 	}{
 		{
-			name:      "retrieve_err",
+			name:      "retrieve_unsupported_location",
 			locations: []string{"mock:", "not_supported:"},
 			parserProvider: map[string]configmapprovider.Provider{
 				"mock": &mockProvider{},
@@ -121,16 +101,6 @@ func TestConfigProvider_Errors(t *testing.T) {
 			parserProvider: map[string]configmapprovider.Provider{
 				"mock": &mockProvider{},
 				"err":  &mockProvider{errR: errors.New("retrieve_err")},
-			},
-			configUnmarshaler: configunmarshaler.NewDefault(),
-			expectNewErr:      true,
-		},
-		{
-			name:      "get_err",
-			locations: []string{"mock:", "err:"},
-			parserProvider: map[string]configmapprovider.Provider{
-				"mock": &mockProvider{},
-				"err":  &mockProvider{ret: &fakeRetrieved{errG: errors.New("retrieve_err")}},
 			},
 			configUnmarshaler: configunmarshaler.NewDefault(),
 			expectNewErr:      true,
@@ -174,7 +144,7 @@ func TestConfigProvider_Errors(t *testing.T) {
 				require.NoError(t, err)
 				return map[string]configmapprovider.Provider{
 					"mock": &mockProvider{},
-					"err":  &mockProvider{ret: &fakeRetrieved{retM: cfgMap, errW: errors.New("watch_err")}},
+					"err":  &mockProvider{retM: cfgMap, errW: errors.New("watch_err")},
 				}
 			}(),
 			configUnmarshaler: configunmarshaler.NewDefault(),
@@ -188,7 +158,7 @@ func TestConfigProvider_Errors(t *testing.T) {
 				require.NoError(t, err)
 				return map[string]configmapprovider.Provider{
 					"mock": &mockProvider{},
-					"err":  &mockProvider{ret: &fakeRetrieved{retM: cfgMap, errC: errors.New("close_err")}},
+					"err":  &mockProvider{retM: cfgMap, errC: errors.New("close_err")},
 				}
 			}(),
 			configUnmarshaler: configunmarshaler.NewDefault(),
@@ -229,7 +199,7 @@ func TestConfigProvider(t *testing.T) {
 		// Use fakeRetrieved with nil errors to have Watchable interface implemented.
 		cfgMap, err := configtest.LoadConfigMap(path.Join("testdata", "otelcol-nop.yaml"))
 		require.NoError(t, err)
-		return &mockProvider{ret: &fakeRetrieved{retM: cfgMap}}
+		return &mockProvider{retM: cfgMap}
 	}()
 
 	cfgW := NewConfigProvider(
@@ -288,7 +258,7 @@ func TestConfigProvider_ShutdownClosesWatch(t *testing.T) {
 		// Use fakeRetrieved with nil errors to have Watchable interface implemented.
 		cfgMap, err := configtest.LoadConfigMap(path.Join("testdata", "otelcol-nop.yaml"))
 		require.NoError(t, err)
-		return &mockProvider{ret: &fakeRetrieved{retM: cfgMap, errW: configsource.ErrSessionClosed}}
+		return &mockProvider{retM: cfgMap, errW: configsource.ErrSessionClosed}
 	}()
 
 	watcherWG := sync.WaitGroup{}