+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Deprecate `configtelemetry.Level.Set()` (#4700)
 - Remove support to some arches and platforms from `ocb` (opentelemetry-collector-builder) (#4710)
 - Remove deprecated legacy path ("v1/trace") support for otlp http receiver (#4720)
+- Replace configmapprovider.Retrieved with returning config.Map and StopWatchFunc (#4721)
 
 ## 💡 Enhancements 💡